refactor(launch): split TUI launcher by concern (Phase 8)
Implements Phase 8 of the Rust structure refactor roadmap. Splits the
1254-LOC `src/launch.rs` into a thin entrypoint + four concern files.
Final `mod.rs` is 57 LOC — only `run_launch` orchestration, terminal
setup/teardown, and the main event loop.

Split:

- `mod.rs` (57 LOC, was 1254) — `pub fn run_launch` entrypoint,
  `TerminalGuard`, `enable_raw_mode`/`EnterAlternateScreen` plumbing,
  the outer event loop, and `pub use` re-exports of `LaunchStage`,
  `LaunchState`, `WorkspaceChoice`.
- `state.rs` (470 LOC) — `LaunchStage`, `WorkspaceChoice`,
  `LaunchState` + its `new` / `selected_workspace_name` /
  `filtered_agents` methods, plus the `configured_agents` and
  `global_mounts` helpers. Co-locates 10 state-level tests,
  including all 6 Phase 0 agent-filter-composition tests from #140
  and the test helpers `agent_source_stub` / `workspace_with_allowed`.
- `input.rs` (110 LOC) — `pub(super) enum EventOutcome` +
  `pub(super) fn handle_event`. Extracted from the inline match
  block inside `run_launch`'s loop. The extraction is the only
  non-verbatim piece of this PR: `break Ok/Err(...)` inside arms
  become `return EventOutcome::Exit(Ok/Err(...))`, and `?` calls
  inside arms expand to explicit match-destructure into
  `EventOutcome::Exit(Err(e))`. Wrapped expressions, key bindings,
  and state mutations are byte-identical to the original.
- `preview.rs` (188 LOC) — `resolve_selected_workspace` and
  `build_agent_detail_lines`. Co-locates the
  `agent_preview_includes_selector_scoped_global_mounts` test.
- `render.rs` (496 LOC) — every draw fn, plus `render_banner`,
  `centered_rect`, `footer_text`, `BANNER_HEIGHT`, the `mod colors`
  palette submodule, and the `footer_text_matches_stage_behavior`
  test.

Verbatim moves for state, preview, render. Only `input.rs` has the
small contained event-handler extraction the plan's Phase 8 names
(`input.rs` is explicitly "event handling and stage transitions").

Test Preservation Policy: all 427 tests preserved, co-located with
the module containing the function they exercise. Test count
unchanged. No test deleted, no test weakened.

Visibility:

- `run_launch` stays `pub`. `LaunchStage` / `LaunchState` /
  `WorkspaceChoice` stay `pub` via `pub use` in `mod.rs` so
  external callers in `src/app/mod.rs` keep their imports.
- `mod colors` bumps from `mod` to `pub(super)` so `preview.rs`
  can use the error/label color tokens when building the detail
  lines. This is the single visibility widening.
- All other new items are `pub(super)` (reachable from sibling
  files under `launch/`, invisible elsewhere).

Zero call sites outside `src/launch/` required edits. The only
external reference (`src/app/mod.rs` calling `launch::run_launch`)
compiles unchanged.

Important constraint honored (per Phase 8 plan): no preview-workspace
type was introduced that shadows `workspace::ResolvedWorkspace`.
`preview.rs` uses the canonical `ResolvedWorkspace` type directly.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 427 tests run: 427 passed, 0 skipped (unchanged)
